sizelu

+++ hello_world/6-size.c
@@ -7,10 +7,10 @@
 
 int main(void)
 {
-    printf("char size: %lu bytes\n", (unsigned long)sizeof(char));
-    printf("int size: %lu bytes\n", (unsigned long)sizeof(int));
-    printf("long int size: %lu bytes\n", (unsigned long)sizeof(long int));
-    printf("long long int size: %lu bytes\n", (unsigned long)sizeof(long long int));
-    printf("float size: %lu bytes\n", (unsigned long)sizeof(float));
+    printf("char size: %lu bytes\n", sizeof(char));
+    printf("int size: %lu bytes\n", sizeof(int));
+    printf("long int size: %lu bytes\n", sizeof(long));
+    printf("long long int size: %lu bytes\n", sizeof(long long int));
+    printf("float size: %lu bytes\n", sizeof(float));
     return (0);
 }
